[CHORE] change directory
