+++ lib/presentation/recipe_screen/recipe_screen.dart
@@ -4,8 +4,10 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_project_orm_janggo/data/db/like_hive/like_item.dart';
+import 'package:flutter_project_orm_janggo/presentation/ads/google_ads_helper.dart';
 import 'package:flutter_project_orm_janggo/presentation/recipe_screen/recipe_view_model.dart';
 import 'package:go_router/go_router.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:provider/provider.dart';
 
 class RecipeScreen extends StatefulWidget {
@@ -27,17 +29,59 @@ class _RecipeScreenState extends State<RecipeScreen> {
   // 현재 페이지의 index
   int _currentPage = 0;
 
+  // initialize google mobile ad sdk
+  Future<InitializationStatus> _initGoogleMobileAds() {
+    return MobileAds.instance.initialize();
+  }
+
+  // Add _interstitialAd
+  InterstitialAd? _interstitialAd;
+
+  // Implement _loadInterstitialAd()
+  void _loadInterstitialAd() {
+    InterstitialAd.load(
+      adUnitId: AdHelper.interstitialAdUnitId,
+      request: AdRequest(),
+      adLoadCallback: InterstitialAdLoadCallback(
+        onAdLoaded: (ad) {
+          print('------------ad------------');
+          ad.fullScreenContentCallback = FullScreenContentCallback(
+            onAdDismissedFullScreenContent: (ad) {},
+          );
+
+          setState(() {
+            _interstitialAd = ad;
+          });
+        },
+        onAdFailedToLoad: (err) {
+          print('Failed to load an interstitial ad: ${err.message}');
+        },
+      ),
+    );
+  }
+
   @override
   void initState() {
     super.initState();
+    _initGoogleMobileAds();
+    _loadInterstitialAd();
     final ingredients = widget.ingredients;
     context.read<RecipeViewModel>().getRecipe(ingredients);
+  }
+
+  // add for ad
+  @override
+  void dispose() {
+    _interstitialAd?.dispose();
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     final viewModel = context.watch<RecipeViewModel>();
     final state = viewModel.state;
+
+    _interstitialAd?.show();
 
     if (state.recipe.isNotEmpty) {
       viewModel.getPicture(state.recipe);
